+++ config/initializers/honeybadger.rb
@@ -1,25 +1,26 @@
 # Can be used to implement more programatic error handling
 # https://docs.honeybadger.io/lib/ruby/getting-started/ignoring-errors.html#ignore-programmatically
 
-MESSAGE_FINGERPRINTS = {
-  "SuspendedError" => "banned",
-  "Rack::Timeout::RequestTimeoutException" => "rack_timeout",
-  "Rack::Timeout::RequestTimeoutError" => "rack_timeout",
-  "PG::QueryCanceled" => "pg_query_canceled"
-}.freeze
-
-COMPONENT_FINGERPRINTS = {
-  "internal" => "internal"
-}.freeze
-
-HONEYBADGER_EXCEPTIONS_TO_IGNORE = [
-  ActiveRecord::QueryCanceled,
-  ActiveRecord::RecordNotFound,
-  Pundit::NotAuthorizedError,
-  RateLimitChecker::LimitReached,
-].freeze
-
+# rubocop:disable Metrics/BlockLength
 Rails.application.reloader.to_prepare do
+  message_fingerprints = {
+    "SuspendedError" => "banned",
+    "Rack::Timeout::RequestTimeoutException" => "rack_timeout",
+    "Rack::Timeout::RequestTimeoutError" => "rack_timeout",
+    "PG::QueryCanceled" => "pg_query_canceled"
+  }
+
+  component_fingerprints = {
+    "internal" => "internal"
+  }
+
+  honeybadger_exceptions_to_ignore = [
+    ActiveRecord::QueryCanceled,
+    ActiveRecord::RecordNotFound,
+    Pundit::NotAuthorizedError,
+    RateLimitChecker::LimitReached,
+  ]
+
   # https://docs.honeybadger.io/lib/ruby/gem-reference/configuration.html
   Honeybadger.configure do |config|
     config.env = "#{ApplicationConfig['APP_DOMAIN']}-#{Rails.env}"
@@ -32,7 +33,7 @@ Rails.application.reloader.to_prepare do
     # Logging allows us to fill in gaps if we need to when errors get discarded.
     config.send_data_at_exit = false
 
-    config.exceptions.ignore += HONEYBADGER_EXCEPTIONS_TO_IGNORE
+    config.exceptions.ignore += honeybadger_exceptions_to_ignore
     config.request.filter_keys += %w[authorization]
     config.sidekiq.attempt_threshold = 10
     config.breadcrumbs.enabled = true
@@ -40,15 +41,16 @@ Rails.application.reloader.to_prepare do
     config.before_notify do |notice|
       notice.fingerprint = if notice.error_message&.include?("SIGTERM") && notice.component&.include?("feeds_import")
                              notice.error_message
-                           elsif (msg_key = MESSAGE_FINGERPRINTS.keys.detect do |k, _v|
+                           elsif (msg_key = message_fingerprints.keys.detect do |k, _v|
                                     notice.error_message&.include?(k)
                                   end)
-                             MESSAGE_FINGERPRINTS[msg_key]
-                           elsif (cmp_key = COMPONENT_FINGERPRINTS.keys.detect do |k, _v|
+                             message_fingerprints[msg_key]
+                           elsif (cmp_key = component_fingerprints.keys.detect do |k, _v|
                                     notice.component&.include?(k)
                                   end)
-                             COMPONENT_FINGERPRINTS[cmp_key]
+                             component_fingerprints[cmp_key]
                            end
     end
   end
 end
+# rubocop:enable Metrics/BlockLength
