+++ spec/initializers/honeybadger_spec.rb
@@ -1,7 +1,12 @@
 require "rails_helper"
 
 describe Honeybadger do
-  MESSAGE_FINGERPRINTS.each do |error_key, fingerprint|
+  {
+    "SuspendedError" => "banned",
+    "Rack::Timeout::RequestTimeoutException" => "rack_timeout",
+    "Rack::Timeout::RequestTimeoutError" => "rack_timeout",
+    "PG::QueryCanceled" => "pg_query_canceled"
+  }.each do |error_key, fingerprint|
     include_examples "#sets_correct_honeybadger_fingerprint", error_key, fingerprint
   end
